# docs/swagger.yaml
definitions:
...
  request.Video:
    properties:
      comments:
        type: integer
      create_by:
        type: string
      description:
        type: string
      likes:
        type: integer
      score:
        type: number
      title:
        minLength: 3
        type: string
      views:
        type: integer
    required:
    - create_by
    - title
    type: object
info:
  contact: {}
paths:
...
  /videos:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all videos
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/request.Video'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List all videos
      tags:
      - videos
    post:
      consumes:
      - application/json
      description: Create a new video in the system
      parameters:
      - description: Video object
        in: body
        name: video
        required: true
        schema:
          $ref: '#/definitions/request.Video'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/request.Video'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new video
      tags:
      - videos
  /videos/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing video
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid video ID
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete a video
      tags:
      - videos
    get:
      consumes:
      - application/json
      description: Get details of a specific video
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/request.Video'
        "400":
          description: Invalid video ID
          schema:
            type: string
        "404":
          description: Video not found
          schema:
            type: string
      summary: Get a video by ID
      tags:
      - videos
    put:
      consumes:
      - application/json
      description: Update an existing video's information
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated video object
        in: body
        name: video
        required: true
        schema:
          $ref: '#/definitions/request.Video'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/request.Video'
        "400":
          description: Invalid video ID or data
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update a video
      tags:
      - videos
  /videos/{id}/comments:
    patch:
      consumes:
      - application/json
      description: Change the number of comments for a specific video
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      - description: Step
        in: query
        name: step
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Comments amount changed
          schema:
            type: string
        "400":
          description: Invalid video ID or step
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Change comments amount
      tags:
      - videos
  /videos/{id}/likes:
    patch:
      consumes:
      - application/json
      description: Change the number of likes for a specific video
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      - description: Step
        in: query
        name: step
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Likes amount changed
          schema:
            type: string
        "400":
          description: Invalid video ID or step
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Change likes amount
      tags:
      - videos
  /videos/{id}/views:
    patch:
      consumes:
      - application/json
      description: Change the number of views for a specific video
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      - description: Step
        in: query
        name: step
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Views amount changed
          schema:
            type: string
        "400":
          description: Invalid video ID or step
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Change views amount
      tags:
      - videos
swagger: "2.0"